master_connections/requirements.txt
anthropic
openai
gensim
python-Levenshtein
nltk
requests
joblib
scikit-learn
python-dotenv
sentence-transformers
pandas
